+++ src/main/java/plugin/AnalyserPlugin.java
@@ -1,4 +1,4 @@
-package gruppun.plugin;
+package plugin;
 
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
