+++ src/main/java/tool/designpatterns/DesignPattern.java
@@ -1,4 +1,4 @@
-package base.designpatterns;
+package tool.designpatterns;
 
 /**
  * Instantiation can be done according to the following example. Using curly brackets to enter an
