+++ src/main/java/tool/designpatterns/Pattern.java
@@ -1,4 +1,4 @@
-package base.designpatterns;
+package tool.designpatterns;
 
 /**
  * An enum that contains all patterns we are able to identify and verify. If a new pattern has been
