+++ src/main/java/tool/designpatterns/verifiers/singleclassverifiers/SingleClassGrouper.java
@@ -1,4 +1,4 @@
-package base.designpatterns.verifiers;
+package tool.designpatterns.verifiers.singleclassverifiers;
 
 import java.util.Iterator;
 import java.util.List;
@@ -6,14 +6,15 @@ import java.util.Map;
 
 import com.github.javaparser.ast.CompilationUnit;
 
-import base.designpatterns.Pattern;
-import base.designpatterns.verifiers.singleclassverifiers.IPatternVerifier;
-import base.util.Feedback;
+import tool.designpatterns.Pattern;
+import tool.designpatterns.verifiers.IPatternGrouper;
+import tool.designpatterns.verifiers.IPatternVerifier;
+import tool.util.Feedback;
 
 /**
  * A class that will verify a PatternGroup that only has one pattern (i.e. single class patterns)
  */
-public class SingleClassVerifier implements IPatternGroupVerifier {
+public class SingleClassGrouper implements IPatternGrouper {
 
     private final transient IPatternVerifier verifier;
 
@@ -22,7 +23,7 @@ public class SingleClassVerifier implements IPatternGroupVerifier {
      *
      * @param verifier the verifier to use when verifying classes.
      */
-    public SingleClassVerifier(IPatternVerifier verifier) {
+    public SingleClassGrouper(IPatternVerifier verifier) {
         this.verifier = verifier;
     }
