+++ src/main/java/tool/designpatterns/verifiers/singleclassverifiers/immutable/ImmutableVerifier.java
@@ -1,4 +1,4 @@
-package base.designpatterns.verifiers.singleclassverifiers;
+package tool.designpatterns.verifiers.singleclassverifiers.immutable;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -22,7 +22,8 @@ import com.github.javaparser.ast.expr.VariableDeclarationExpr;
 import com.github.javaparser.ast.stmt.BlockStmt;
 import com.github.javaparser.ast.stmt.Statement;
 
-import base.util.Feedback;
+import tool.designpatterns.verifiers.IPatternVerifier;
+import tool.util.Feedback;
 
 /**
  * A verifier for the immutable pattern.
