+++ src/main/java/tool/util/Feedback.java
@@ -1,4 +1,4 @@
-package base.util;
+package tool.util;
 
 /**
  * Simple class to store and manage feedback to the user.
